backend/requirements.txt
asgiref==3.8.1
Django==4.2.16
django-cors-headers==4.6.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==10.3.0
psycopg2-binary==2.9.9
sqlparse==0.5.2
typing_extensions==4.12.2
